% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcontributor-package.R
\docType{package}
\name{allcontributor-package}
\alias{allcontributor}
\alias{allcontributor-package}
\title{allcontributor: Acknowledge all contributors in your 'README'}
\description{
Acknowledge all contributors in your 'README'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/allcontributor}
  \item Report bugs at \url{https://github.com/mpadge/allcontributor/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
