% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcontributor-package.R
\docType{package}
\name{allcontributors-package}
\alias{allcontributors}
\alias{allcontributors-package}
\title{allcontributors: Acknowledge all Contributors to a Project}
\description{
Acknowledge all contributors to a project via a single function call. The function appends to a 'README' or other specified file(s) a table with names of all individuals who contributed via code or repository issues. The package also includes several additional functions to extract and quantify contributions to any repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/allcontributors}
  \item Report bugs at \url{https://github.com/ropensci/allcontributors/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Chris Hartgerink
}

Other contributors:
\itemize{
  \item Maëlle Salmon [contributor]
}

}
\keyword{internal}
