% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_gh_contrib_issue}
\alias{get_gh_contrib_issue}
\title{get_gh_contrib_issue}
\usage{
get_gh_contrib_issue(org, repo)
}
\arguments{
\item{org}{Github organisation name for repository}

\item{repo}{Repository within \code{org} for which contributors are to be
extracted}
}
\value{
Character vector of github logins for all contributors listed in
current issue
}
\description{
Extract contributors currently listed on an "All Contributions" issue in a
github repository. This is much easier with the REST API than via graphql.
}
