% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_gh_issue_titles}
\alias{get_gh_issue_titles}
\title{get_gh_issue_titles}
\usage{
get_gh_issue_titles(org, repo)
}
\arguments{
\item{org}{Github organisation name for repository}

\item{repo}{Repository within \code{org} for which contributors are to be
extracted}
}
\value{
\code{data.frame} with one column of issue numbers, and one column of
issue titles.
}
\description{
Extract titles and numbers of all issues associated with a nominated
repository
}
\examples{
\dontrun{
get_gh_issue_titles (org = "ropenscilabs", repo = "allcontributors")
}
}
\seealso{
Other github: 
\code{\link{get_gh_code_contributors}()},
\code{\link{get_gh_contrib_issue}()},
\code{\link{get_gh_issue_people}()}
}
\concept{github}
