% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-contributors.R
\name{add_contributors}
\alias{add_contributors}
\title{add_contributors}
\usage{
add_contributors(ncols = 7, alphabetical = FALSE)
}
\arguments{
\item{ncols}{Number of columns for contributors in 'README'}

\item{alphabetical}{If \code{TRUE}, order contributors alphabetically, otherwise
order by decreasing numbers of contributions.}
}
\value{
Nothing.
}
\description{
Add contributors to README.Rmd
}
