% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcontributor-package.R
\docType{package}
\name{allcontributors-package}
\alias{allcontributors}
\alias{allcontributors-package}
\title{allcontributors: Acknowledge all contributors in your 'README'}
\description{
Acknowledge all contributors in your 'README'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/allcontributors}
  \item Report bugs at \url{https://github.com/ropenscilabs/allcontributors/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
